requirements.txt
streamlit==1.37.1
pydrive2==1.20.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
xlrd==2.0.1
lxml==5.2.2
beautifulsoup4==4.12.3
rapidfuzz==3.9.6
unidecode==1.3.8